### 2023-01-11 #571 @mikemahoney218
waywiser: Ergonomic Methods for Assessing Spatial Models
Submitting Author Name: Mike Mahoney
Submitting Author Github Handle: @mikemahoney218
Repository:  https://github.com/mikemahoney218/waywiser
Version submitted:
Submission type: Stats
Badge grade: silver
Reviewers: @becarioprecario, @jakub_nowosad, @nowosad
Version accepted: TBD
---

```
Package: waywiser
Title: Ergonomic Methods for Assessing Spatial Models
Version: 0.2.0.9000
    person("Michael", "Mahoney", , "mike.mahoney.218@gmail.com", role = c("aut", "cre"),
           comment = c(ORCID = "0000-0003-2402-304X")),
    person("Lucas", "Johnson", , "lucas.k.johnson03@gmail.com", role = c("ctb"),
    person("RStudio", role = c("cph", "fnd"))
  )
Description: Assessing predictive models of spatial data can be challenging,
    both because these models are typically built for extrapolating outside the
    original region represented by training data and due to potential spatially
    structured errors, with "hot spots" of higher than expected error
    clustered geographically due to spatial structure in the underlying
    data. Methods are provided for assessing models fit to spatial data,
    including approaches for measuring the spatial structure of model errors,
    assessing model predictions at multiple spatial scales, and evaluating where
    predictions can be made safely. Methods are particularly useful for models
    fit using the 'tidymodels' framework. Methods include Moran's I
    ('Moran' (1950) <doi:10.2307/2332142>), Geary's C
    ('Geary' (1954) <doi:10.2307/2986645>), Getis-Ord's G
    ('Ord' and 'Getis' (1995) <doi:10.1111/j.1538-4632.1995.tb00912.x>),
    agreement coefficients from 'Ji' and Gallo (2006)
    (<doi: 10.14358/PERS.72.7.823>), agreement metrics from 'Willmott' (1981)
    (<doi: 10.1080/02723646.1981.10642213>) and 'Willmott' 'et' 'al'. (2012)
    (<doi: 10.1002/joc.2419>), an implementation of the area of applicability
    methodology from 'Meyer' and 'Pebesma' (2021)
    (<doi:10.1111/2041-210X.13650>), and an implementation of
    multi-scale assessment as described in 'Riemann' 'et' 'al'. (2010)
    (<doi:10.1016/j.rse.2010.05.010>).
    https://mikemahoney218.github.io/waywiser/
    fields,
    FNN,
    glue,
    hardhat,
    Matrix,
    rsample,
    sf (>= 1.0-0),
    spdep (>= 1.1-9),
    tidyselect,
    yardstick
    applicable,
    caret,
    CAST,
    covr,
    modeldata,
    recipes,
    spatialsample,
    tidymodels,
    tigris,
    vip,
    whisker,
    withr
Config/testthat/parallel: true
Roxygen: list(markdown = TRUE, roclets = c("namespace", "rd", "srr::srr_stats_roclet"))
```

- Please indicate which of our [statistical package categories](https://stats-devguide.ropensci.org/overview.html#overview-categories) this package falls under. (Please check one appropriate box below):

- [x] A pre-submission inquiry has been approved in [issue#565](#565)

Anyone fitting models to spatial data, particularly (but not exclusively) people working within the tidymodels ecosystem. This includes a number of domains, and we've already been using it in our modeling practice.

    - *The first implementation of a novel algorithm*; or
    - *The first implementation within* **R** *of an algorithm which has previously been implemented in other languages or contexts*; or
    - *An improvement on other implementations of similar algorithms in* **R**.

    Please include hyperlinked references to all other relevant software.

The waywiser R package makes it easier to measure the performance of models fit to 2D spatial data by implementing a number of well-established assessment methods in a consistent, ergonomic toolbox; features include new [yardstick](https://yardstick.tidymodels.org/) metrics for measuring agreement and spatial autocorrelation, functions to assess model predictions across multiple scales, and methods to calculate the area of applicability of a model.

Relevant software implementing similar algorithms include [CAST](https://github.com/HannaMeyer/CAST) for `ww_area_of_applicability()`. Several yardstick metrics implemented directly wrap [spdep](https://github.com/r-spatial/spdep/) in a more consistent interface. Willmott's D is also implemented in [hydroGOF](https://github.com/hzambran/hydroGOF). Other functions have (as far as I am aware) not been implemented elsewhere, such as `ww_multi_scale()` which implements the procedure from [Riemann et al 2010](https://doi.org/10.1016/j.rse.2010.05.010), or `ww_agreement_coefficient()` which implements metrics from [Ji and Gallo 2006](https://doi.org/10.14358/PERS.72.7.823).

N/A

Silver

> Have a demonstrated generality of usage beyond one single envisioned use case. Software is frequently developed for one particular use case envisioned by the authors themselves. Generalising the utility of software so that it is readily applicable to other use cases, and satisfactorily documenting such generality of usage, represents another aspect which may be considered sufficient for software to attain a silver grade.

This is the primary aspect which I believe merits the silver status. The waywiser package implements routines which are useful for a wide variety of spatial models and integrates well with the tidymodels ecosystem, making it (hopefully!) of interdisciplinary interest.

Depending on what the editors think, I'd also potentially submit this for gold, based upon the following two aspects:

> Compliance with a good number of standards beyond those identified as minimally necessary. This will require reviewers and authors to agree on identification of both a minimal subset of necessary standards, and a full set of potentially applicable standards. This aspect may be considered fulfilled if at least one quarter of the additional potentially applicable standards have been met, and should definitely be considered fulfilled if more than one half have been met.

> Internal aspects of package structure and design. Many aspects of the internal structure and design of software are too variable to be effectively addressed by standards. Packages which are judged by reviewers to reflect notably excellent design choices, especially in the implementation of core statistical algorithms, may also be considered worthy of a silver grade.

But I'm not familiar enough with the system to know if waywiser is likely to be in compliance with these two aspects, and am comfortable submitting for "silver" status if waywiser does not obviously meet both.

- [ ] I/we have run [`autotest`](https://github.com/ropensci-review-tools/autotest) checks on the package, and ensured no tests fail. (Sorry, both the release and CRAN versions of autotest fail immediately on my machine with internal errors -- that is, from autotest itself and not from my package -- and therefore I have not been able to use it).
