### 2023-01-12 #571 @annakrystalli
Hello @mikemahoney218 ,

 The failures shown on the checks appear to be genuine and are pointing to some issues in the package.
It appears the package fails `rcmdcheck` because of the stuff shown at the bottom of "goodpractice and other checks", with vignettes producing a core dump.

We encourage you to try and reproduce in a clean docker container by cloning the repo and running `rcmdcheck::rcmdcheck()` and see what you get.

With respect to messages from statistical checks by `srr`, we are investigating as we thing those are issues on our end.
