### 2023-01-12 #571 @annakrystalli
BTW @mikemahoney218, regarding testing locally in a docker container, [this section of our devguide](https://devguide.ropensci.org/ci.html?q=docker#testing-using-different-versions-of-r) has some useful links to info (just in case).
