### 2023-01-12 #571 @mpadge
@mikemahoney218 @annakrystalli The `srr` section of the checks has now also been updated. Sorry for any inconvenience. @mikemahoney218 Please call `@ropensci-review-bot check package` to re-generate the checks once you've addressed both the `srr` and the failing `rcmdcheck` issues.
