### 2023-01-12 #571 @mikemahoney218
Hi @annakrystalli & @mpadge -- can I ask if there's more information about your CI server available anywhere? I'm wondering what results you get from `sf::sf_extSoftVersion()` (and `sessionInfo()`), as this issue seems to be local to your CI setup.

I can't reproduce the issue on CRAN:
![image](https://user-images.githubusercontent.com/38229299/212073242-8f15efe6-6d81-4a0d-a3b7-c7311360d7cd.png)
(Currently at https://win-builder.r-project.org/CY6Z7In5rrks/, expect that link will break after 1/15 though)

On CI:
https://github.com/mikemahoney218/waywiser/pull/15

Or locally on Docker:
![image](https://user-images.githubusercontent.com/38229299/212079657-46d5bc56-5900-4a4c-906a-d70daa4d8aa0.png)

I notice that your link suggests the R-Hub docker images; it has been my experience that R-Hub has not been able to install most spatial software for a few years now. I checked using the rocker images, via the command:
```bash
docker run --rm -ti -v "$(pwd)":/home/rstudio rocker/geospatial R
```
(The volume attaches my code folder as the home directory in order to check the package.)

So it seems like I'm not able to reproduce this issue across a variety of environments.
