### 2023-01-12 #571 @mpadge
@mikemahoney218 It's our own docker image used specifically for package checks. Current version gives this:

``` r
sf::sf_extSoftVersion()
#>           GEOS           GDAL         proj.4 GDAL_with_GEOS     USE_PROJ_H
#>       "3.8.0"          "3.0.4"        "6.3.1"         "true"         "true"
#>           PROJ
#>        "6.3.1"
```

... but i can confirm that the issue is directly caused by `sf`, and not your package. Wrong linkage with compiled version of GEOS. I'll ping here once we've fixed that up and can run the check again.  That might take a while, so in the meantime please ignore those fails and accept our aplogies. Thanks.
