### 2023-01-12 6c57cc8
Add missing standards (#15)
* Add missing standards

* Update snapshots
