### 2023-01-12 #571 @mikemahoney218
Hi @annakrystalli ! I'm not sure how long the checks _should_ take, but we're a bit past two hours now. I believe I've fixed the srr issue, and it sounds like fixing the CI system may take a while, but the package works on non-rOpenSci systems.
