+++ pkgdown.yml
@@ -1,8 +1,11 @@
-pandoc: 2.14.2
-pkgdown: 2.0.6
+pandoc: 2.19.2
+pkgdown: 2.0.7
 pkgdown_sha: ~
-articles: {}
-last_built: 2022-10-23T04:24Z
+articles:
+  multi-scale-assessment: multi-scale-assessment.html
+  residual-autocorrelation: residual-autocorrelation.html
+  waywiser: waywiser.html
+last_built: 2023-01-12T23:34Z
 urls:
   reference: https://mikemahoney218.github.io/waywiser/reference
   article: https://mikemahoney218.github.io/waywiser/articles
