+++ pkgdown.yml
@@ -5,7 +5,7 @@ articles:
   multi-scale-assessment: multi-scale-assessment.html
   residual-autocorrelation: residual-autocorrelation.html
   waywiser: waywiser.html
-last_built: 2023-01-14T14:19Z
+last_built: 2023-01-14T16:00Z
 urls:
   reference: https://mikemahoney218.github.io/waywiser/reference
   article: https://mikemahoney218.github.io/waywiser/articles
