### 2023-01-25 #571 @becarioprecario
- [ ] **Examples:** (that run successfully locally) for all exported functions. NOTE FROM REVIEWER: Some of the exported functions seem to be internal functions (see my comment below).

---

Thanks for contributing the package to the community! I hope that the following comments can help you "fine tune" some parts of the package:

* Vignettes: I have not been able to find any vignette in the package. There is an example in the README.md file that could be moved to a vignette.

* Manual pages: I think that the examples included in the package include cases studies for all main functions in the package. For example, file `global_geary_c.Rd` includes an example in which function `ww_global_geary()` is exemplified. All the other functions described in the manual page are internal functions called from `ww_global_geary()`. At least, this is my feeling after reading the Rd files.

* Related to my previous comment, it may be good to keep some functions as internal and not exported in the NAMESPACE file. For example, functions to compute p-values and possibly others.

* In one of the examples I get some code shown that, I guess, should not be shown:

```
ww_gl_> ## Don't show:
ww_gl_> if (rlang::is_installed("sfdep")) (if (getRversion() >= "3.4") withAutoprint else force)({ # examplesIf
ww_gl_+ ## End(Don't show)
```

* Also, when running `example(ww_global_geary, run.dontrun=TRUE)` the example does not seem to run at all. Please, could the author check this and, possibly, all the other examples to make sure that `\dontshow` and `\dontrun` marks are properly used.

* I do not think that it is a good practice to write examples conditional on some package being installed. In particular, the `sfdep` package. PLease, simply make it a dependence to make sure that the package is installed before your package and that examples can be run. Otherwise the examples cannot be run... this is why I was having the behaviour in my previous comment (which I left because it was something that I was not expecting at all).

* I do not think that the package fullfils the [**Packaging guidelines**](https://devguide.ropensci.org/building.html#building) as there are no vignettes in the package.

* When I run, for example, `ww_global_geary(guerry_modeled, crime_pers, predictions)` it is not clear to me what to do with the output. It provides a `tibble` with the estimate and a p-value, as well as a geometry in both lines. What is the geometry for?

* Also, is there a way of getting a summary of the test? I mean, in a similar way as standard functions in the `spdep` package for Geary's c test, etc.
