### 2023-01-25 #571 @mikemahoney218
@becarioprecario , thank you very much for your thoughtful review!

If I may, can I ask which version of the package you were reviewing? My intention was to submit the [current development version of the package](https://github.com/mikemahoney218/waywiser), which has several changes from the version currently on CRAN which fix elements of your review; notably, the `ww_global_geary()` function you mentioned no longer exists, `sfdep` is no longer in any package files at all, and there are [three vignettes](https://mikemahoney218.github.io/waywiser/articles/).

I apologize for the confusion here; I know it's not best practice for a submission package to already be on CRAN, but this package has expanded significantly in scope from the first CRAN release to the point that I thought it was worth an rOpenSci submission.

cc: @annakrystalli @Nowosad -- apologies if I wasn't clear enough which version was submitted for review; it's meant to be the [GitHub version](https://github.com/mikemahoney218/waywiser).
