### 2023-01-26 #571 @becarioprecario
Yes, I did review the package on CRAN. I will go through the GitHub version later and see if some of my comments still apply.

El 25 ene 2023, a las 23:45, Michael Mahoney ***@***.***> escribió:

@becarioprecario<https://github.com/becarioprecario> , thank you very much for your thoughtful review!

If I may, can I ask which version of the package you were reviewing? My intention was to submit the current development version of the package<https://github.com/mikemahoney218/waywiser>, which has several changes from the version currently on CRAN which fix elements of your review; notably, the ww_global_geary() function you mentioned no longer exists, sfdep is no longer in any package files at all, and there are three vignettes<https://mikemahoney218.github.io/waywiser/articles/>.

I apologize for the confusion here; I know it's not best practice for a submission package to already be on CRAN, but this package has expanded significantly in scope from the first CRAN release to the point that I thought it was worth an rOpenSci submission.

cc: @annakrystalli<https://github.com/annakrystalli> @Nowosad<https://github.com/Nowosad> -- apologies if I wasn't clear enough which version was submitted for review; it's meant to be the GitHub version<https://github.com/mikemahoney218/waywiser>.
