### 2023-01-26 #571 @Paula-Moraga
Many thanks @becarioprecario for your quick review. We really appreciate your time and effort. We apologize for the confusion on the version to be reviewed. We are working to better [clarify the guidelines](https://github.com/ropensci/dev_guide/issues/594) to make it clear that the review should be about the GitHub version and not the CRAN version of the package. Also apologies to @mikemahoney218 for the confusion and many thanks for your clarification about your submission.
