### 2023-01-27 #571 @becarioprecario
I have now reviewed package version 0.2.0.9000 from GitHub.

- [X] **Examples:** (that run successfully locally) for all exported functions. NOTE FROM REVIEWER: Some of the exported functions seem to be internal functions (see my comment below).

---

Thanks for contributing the package to the community! I hope that the following comments can help you "fine tune" some parts of the package:

* Vignettes: The vignettes seem to describe major functionalities in the package. I wonder whether in the main vignette `waywiser` the author could include a summary of all the capabilities of the R package (e.g., a table with a list of the available statistics implemented: name, descrption and, maybe, reference).

* Vignette `residual-autocorrelation` I would use title `Local Moran` rather than `Morans I` in the label of the legend as the local Moran is what appears in the plot.

* Manual pages: I think that the examples included in the package include cases studies for all main functions in the package. For example, file `global_geary_c.Rd` includes an example in which function `ww_global_geary()` is exemplified. All the other functions described in the manual page are internal functions called from `ww_global_geary()`. At least, this is my feeling after reading the Rd files.

* Related to my previous comment, it may be good to keep some functions as internal and not exported in the NAMESPACE file. For example, functions to compute p-values and possibly others.

* When I run, for example, `ww_global_geary(guerry_modeled, crime_pers, predictions)` it is not clear to me what to do with the output. It provides a `tibble` with the estimate and a p-value, as well as a geometry in both lines. What is the geometry for?

* Also, is there a way of getting a summary of the test? I mean, in a similar way as standard functions in the `spdep` package for Geary's c test, etc.
