+++ R/area_of_applicability.R
@@ -152,6 +152,14 @@
 #' aoa <- ww_area_of_applicability(y ~ ., train, test, importance = importance)
 #' predict(aoa, test)
 #'
+#' # Equivalent methods for calculating AOA:
+#' ww_area_of_applicability(train[2:11], test[2:11], importance)
+#' ww_area_of_applicability(
+#'   as.matrix(train[2:11]),
+#'   as.matrix(test[2:11]),
+#'   importance
+#' )
+#'
 #' @srrstats {G1.0} Reference for this method:
 #' @references
 #' H. Meyer and E. Pebesma. 2021. "Predicting into unknown space? Estimating
