+++ R/multi_scale.R
@@ -96,7 +96,7 @@
 #'
 #' @srrstats {G1.0} Reference for these methods:
 #' @references
-#' Riemann, R., Wilson, B. T., Lister, A., and Parks, S. "An effective
+#' Riemann, R., Wilson, B. T., Lister, A., and Parks, S. (2010). "An effective
 #' assessment protocol for continuous geospatial datasets of forest
 #' characteristics using USFS Forest Inventory and Analysis (FIA) data."
 #' Remote Sensing of Environment 114(10), pp 2337-2352,
