+++ R/willmott_d.R
@@ -57,6 +57,14 @@
 #' ww_systematic_rmse_vec(x, y)
 #' ww_unsystematic_rmse_vec(x, y)
 #'
+#' example_df <- data.frame(x = x, y = y)
+#' ww_willmott_d(example_df, x, y)
+#' ww_willmott_dr(example_df, x, y)
+#' ww_systematic_mse(example_df, x, y)
+#' ww_unsystematic_mse(example_df, x, y)
+#' ww_systematic_rmse(example_df, x, y)
+#' ww_unsystematic_rmse(example_df, x, y)
+#'
 #' @srrstats {G1.0} Reference for these methods:
 #' @references
 #' Willmott, C. J. 1981. "On the Validation of Models". Physical Geography 2(2),
