+++ R/global_geary.R
@@ -57,6 +57,19 @@
 #' ww_global_geary_c(guerry_model, Crm_prs, predictions)
 #' ww_global_geary_pvalue(guerry_model, Crm_prs, predictions)
 #'
+#' wt <- ww_build_weights(guerry_model)
+#'
+#' ww_global_geary_c_vec(
+#'   guerry_model$Crm_prs,
+#'   guerry_model$predictions,
+#'   wt = wt
+#' )
+#' ww_global_geary_pvalue_vec(
+#'   guerry_model$Crm_prs,
+#'   guerry_model$predictions,
+#'   wt = wt
+#' )
+#'
 #' @references
 #' Geary, R. C. (1954). "The Contiguity Ratio and Statistical Mapping". The
 #' Incorporated Statistician. 5 (3): 115–145. doi:10.2307/2986645.
