+++ R/local_geary.R
@@ -52,6 +52,19 @@
 #' ww_local_geary_c(guerry_model, Crm_prs, predictions)
 #' ww_local_geary_pvalue(guerry_model, Crm_prs, predictions)
 #'
+#' wt <- ww_build_weights(guerry_model)
+#'
+#' ww_local_geary_c_vec(
+#'   guerry_model$Crm_prs,
+#'   guerry_model$predictions,
+#'   wt = wt
+#' )
+#' ww_local_geary_pvalue_vec(
+#'   guerry_model$Crm_prs,
+#'   guerry_model$predictions,
+#'   wt = wt
+#' )
+#'
 #' @references
 #' Anselin, L. 1995. Local indicators of spatial association, Geographical
 #' Analysis, 27, pp 93–115. doi: 10.1111/j.1538-4632.1995.tb00338.x.
