+++ man/global_moran_i.Rd
@@ -74,6 +74,19 @@ guerry_model$predictions <- predict(guerry_lm, guerry_model)
 ww_global_moran_i(guerry_model, Crm_prs, predictions)
 ww_global_moran_pvalue(guerry_model, Crm_prs, predictions)
 
+wt <- ww_build_weights(guerry_model)
+
+ww_global_moran_i_vec(
+  guerry_model$Crm_prs,
+  guerry_model$predictions,
+  wt = wt
+)
+ww_global_moran_pvalue_vec(
+  guerry_model$Crm_prs,
+  guerry_model$predictions,
+  wt = wt
+)
+
 }
 \references{
 Moran, P.A.P. (1950). "Notes on Continuous Stochastic Phenomena." Biometrika,

[…]
