### 2023-01-30 #571 @mikemahoney218
Hi all!

I've started responding to review comments in a [new branch](https://github.com/mikemahoney218/waywiser/pull/16). I'm intending to wait to merge this until all reviews are in, to make sure I'm not asking Jakub to review a moving target :smile:

That said, I wanted to highlight that a _lot_ of the tests (235 in total) were broken by yesterday's dplyr 1.1.0 release; these are fixed on the reviewer_comments branch, but may be a bit distracting if you review the main branch with the newest dplyr version installed.

In my view, the only test fix that's worth noting is that the `ww_local_*` functions rely on using `dplyr::summarise()` to return multiple rows, which returns a warning in dplyr 1.1.0. I've opened [an issue in yardstick](https://github.com/tidymodels/yardstick/issues/355) to see if this can be fixed upstream or if I need to work around the issue in waywiser itself, and for the time being simply ignored the warning in tests so that CI is a bit more useful, but wanted to clarify that I'm not planning on ignoring the warning long-term.
