+++ .github/workflows/R-CMD-check.yaml
@@ -31,7 +31,8 @@ jobs:
           - {os: ubuntu-latest,   r: 'release'}
           - {os: ubuntu-latest,   r: 'oldrel-1'}
           - {os: ubuntu-latest,   r: 'oldrel-2'}
-          - {os: ubuntu-latest,   r: 'oldrel-3'}
+          # Until dplyr::summarize is swapped for dplyr::reframe
+#          - {os: ubuntu-latest,   r: 'oldrel-3'}
           # Uncomment this in 2023 (once current R is >= 4.3)
           # The package requires R 3.6, I don't really remember why
 #          - {os: ubuntu-latest,   r: 'oldrel-4'}
@@ -53,6 +54,7 @@ jobs:
 
       - uses: r-lib/actions/setup-r-dependencies@v2
         with:
+          cache-version: 2
           extra-packages: any::rcmdcheck, CAST=?ignore-before-r=4.1.0
           needs: check
 
