### 2023-01-31 68416f7
Address reviewer comments (#16)
* Address comments

* Fix examples, spelling

* Update for new dplyr

* CI cache bust
