### 2023-01-31 #571 @mikemahoney218
Sounds good! I've merged the `reviewer_comments` branch, so tests now are updated to (and expect) dplyr 1.1.0.
