### 2023-02-02 #571 @Nowosad
## Package Review

-   ☒ As the reviewer I confirm that there are no [conflicts of interest](https://devguide.ropensci.org/policies.html#coi) for me to review this work (if you are unsure whether you are in conflict, please speak to your editor *before* starting your review).

#### Documentation

-   ☒ **A statement of need:** clearly stating problems the software is designed to solve and its target audience in README
-   ☒ **Installation instructions:** for the development version of package and any non-standard dependencies in README
-   ☒ **Vignette(s):** demonstrating major functionality that runs successfully locally
-   ☒ **Function Documentation:** for all exported functions
-   ☒ **Examples:** (that run successfully locally) for all exported functions
-   ☒ **Community guidelines:** including contribution guidelines in the README or CONTRIBUTING, and DESCRIPTION with `URL`, `BugReports` and `Maintainer` (which may be autogenerated via `Authors@R`).

#### Functionality

-   ☒ **Installation:** Installation succeeds as documented.
-   ☒ **Functionality:** Any functional claims of the software been confirmed.
-   ☒ **Performance:** Any performance claims of the software been confirmed.
-   ☒ **Automated tests:** Unit tests cover essential functions of the package and a reasonable range of inputs and conditions. All tests pass on the local machine.
-   ☒ **Packaging guidelines**: The package conforms to the rOpenSci packaging guidelines.

Estimated hours spent reviewing:

3

-   ☒ Should the author(s) deem it appropriate, I agree to be acknowledged as a package reviewer (“rev” role) in the package DESCRIPTION file.

### Review Comments

In general, I find the package very well written and documented. I only have a few comments/suggestions:

-   waywiser.Rmd: Have you considered explaining the example data first in this vignette?
-   waywiser.Rmd: Could you add a sentence or two better explaining where `n` came from in the Multi-scale model assessment section?; why is this a list (and also why `cellsize` is not a list in multi-scale-assessment.Rmd)?
-   waywiser.Rmd: Area of Applicability: could you expand the description of the `importance` argument in this vignette?
-   waywiser.Rmd: Area of Applicability: I would suggest also showing the result here (it is hard to think about an area of applicability without seeing it first)
-   waywiser.Rmd: I like the Feature Matrix table, however, I am not sure if it should be at the end of this vignette. Have you considered moving it to a standalone vignette (for better visibility)? Also, could you replace DOI codes with DOI urls?
-   residual-autocorrelation.Rmd and multi-scale-assessment.Rmd: What is the reason for using the %&gt;% pipe here? Why not use the native pipe (|&gt;)?
-   residual-autocorrelation.Rmd: “This makes it easy to see what areas are poorly represented by our model” – could you elaborate on this sentence and explain which areas you are talking about?
-   GitHub Actions are mostly broken at the moment (I assume it is due to the recent dplyr changes)
-   Regarding your “p-value” question: I think option 2 is fine.
-   Question: How do you see the future of this package? Do you plan to add any new features (e.g., spatial explainers)?
