+++ R/local_moran.R
@@ -3,6 +3,9 @@
 #' Calculate the local Moran's I statistic for model residuals.
 #' `ww_local_moran_i()` returns the statistic itself, while
 #' `ww_local_moran_pvalue()` returns the associated p value.
+#' These functions are meant to help assess model predictions, for instance by
+#' identifying clusters of higher residuals than expected. For statistical
+#' testing and inference applications, use [spdep::localmoran_perm()] instead.
 #'
 #' These functions can be used for geographic or projected coordinate reference
 #' systems and expect 2D data.
