+++ R/local_geary.R
@@ -3,6 +3,9 @@
 #' Calculate the local Geary's C statistic for model residuals.
 #' `ww_local_geary_c()` returns the statistic itself, while
 #' `ww_local_geary_pvalue()` returns the associated p value.
+#' These functions are meant to help assess model predictions, for instance by
+#' identifying clusters of higher residuals than expected. For statistical
+#' testing and inference applications, use [spdep::localC_perm()] instead.
 #'
 #' These functions can be used for geographic or projected coordinate reference
 #' systems and expect 2D data.
