+++ vignettes/waywiser.Rmd
@@ -30,22 +30,31 @@ things:
    extending this tool to work with tidymodels infrastructure.
 
 This vignette will walk through each of these goals in turn. Before we do that, 
-let's set up the data we'll use in examples:
+let's set up the data we'll use in examples. We'll be using simulated data
+based on [Worldclim](https://www.worldclim.org/) variables; our predictors here
+represent temperature and precipitation values at sampled locations, while, 
+our response represents a virtual species distribution:
 
 ```{r setup}
 library(waywiser)
+set.seed(1107)
+
+worldclim_training <- sample(nrow(worldclim_simulation) * 0.8)
+worldclim_testing <- worldclim_simulation[-worldclim_training, ]
+worldclim_training <- worldclim_simulation[worldclim_training, ]
+
 
 worldclim_model <- lm(
   response ~ bio2 + bio10 + bio13 + bio19, 
-  worldclim_simulation
+  worldclim_training
 )
 
-worldclim_simulation$predictions <- predict(
+worldclim_testing$predictions <- predict(
   worldclim_model,
-  worldclim_simulation
+  worldclim_testing
 )
 
-head(worldclim_simulation)
+head(worldclim_testing)
 ```
 
 ## Yardstick Extensions
@@ -61,14 +70,14 @@ coefficient from [Ji and Gallo 2006](https://doi.org/10.14358/PERS.72.7.823):
 
 ```{r}
 ww_agreement_coefficient(
-  worldclim_simulation,
+  worldclim_testing,
   truth = response,
   estimate = predictions
 )
 
 ww_agreement_coefficient_vec(
-  truth = worldclim_simulation$response,
-  estimate = worldclim_simulation$predictions
+  truth = worldclim_testing$response,
+  estimate = worldclim_testing$predictions
 )
 ```
 
@@ -77,7 +86,7 @@ Some of these additional metrics are implemented by wrapping functions from the
 
 ```{r}
 ww_global_geary_c(
-  worldclim_simulation,
+  worldclim_testing,
   truth = response,
   estimate = predictions
 )
@@ -91,14 +100,14 @@ or a function that will create spatial weights:
 
 ```{r}
 ww_global_geary_c(
-  worldclim_simulation,
+  worldclim_testing,
   truth = response,
   estimate = predictions,
-  wt = ww_build_weights(worldclim_simulation)
+  wt = ww_build_weights(worldclim_testing)
 )
 
 ww_global_geary_c(
-  worldclim_simulation,
+  worldclim_testing,
   truth = response,
   estimate = predictions,
   wt = ww_build_weights
@@ -112,7 +121,7 @@ Because these are yardstick metrics, they can be used with
 yardstick::metric_set(
   ww_agreement_coefficient,
   ww_global_geary_c
-)(worldclim_simulation,
+)(worldclim_testing,
   truth = response,
   estimate = predictions)
 ```
@@ -129,22 +138,35 @@ recommend assessing model predictions aggregated to multiple scales.
 The `ww_multi_scale()` function helps automate this process. The interface for
 this function works similarly to that for yardstick metrics -- you provide your
 data, your true values, and your estimate -- except you also must provide 
-instructions for how to aggregate your data. This can take the form of providing
-information on how to build standardized grids, by passing arguments that will
-be used by `sf::st_make_grid()`, or by passing polygons directly:
+instructions for how to aggregate your data. You can do this by passing 
+arguments that will be used by `sf::st_make_grid()`; for instance, we can use
+the `n` argument to control how many polygons our grid has in the x and y 
+directions. 
+
[…]
