### 2023-02-02 a391ae2
Vignette doesn't need kableExtra anymore
