+++ .github/workflows/R-CMD-check.yaml
@@ -56,11 +56,6 @@ jobs:
           extra-packages: any::rcmdcheck, CAST=?ignore-before-r=4.1.0
           needs: check
 
-      - name: Update dependencies
-        run: |
-          install.packages("purrr", repos = 'https://cloud.r-project.org')
-        shell: Rscript {0}
-
       - uses: r-lib/actions/check-r-package@v2
         env:
           _R_CHECK_FORCE_SUGGESTS_: false
