+++ tests/testthat/test-global_geary.R
@@ -1,5 +1,5 @@
 test_that("Global Geary statistics are stable", {
-
+  trip_dplyr_warning()
   guerry_modeled <- guerry
   guerry_lm <- lm(Crm_prs ~ Litercy, guerry_modeled)
   guerry_modeled$predictions <- predict(guerry_lm, guerry_modeled)
