+++ tests/testthat/test-misc.R
@@ -1,5 +1,5 @@
 test_that("ww_make_point_neighbors is stable", {
-
+  trip_dplyr_warning()
   guerry_pt <- sf::st_geometry(guerry)
   guerry_pt <- sf::st_centroid(guerry_pt)
   expect_snapshot(
@@ -13,7 +13,7 @@ test_that("ww_make_point_neighbors is stable", {
 })
 
 test_that("ww_make_polygon_neighbors is stable", {
-
+  trip_dplyr_warning()
   expect_snapshot(
     ww_make_polygon_neighbors(guerry)
   )
@@ -25,7 +25,7 @@ test_that("ww_make_polygon_neighbors is stable", {
 })
 
 test_that("ww_build_neighbors is stable", {
-
+  trip_dplyr_warning()
   expect_snapshot(
     ww_build_neighbors(guerry)
   )
@@ -53,7 +53,7 @@ test_that("ww_build_neighbors is stable", {
 })
 
 test_that("ww_build_weights is stable", {
-
+  trip_dplyr_warning()
   expect_snapshot(
     ww_build_weights(guerry)
   )
@@ -85,6 +85,7 @@ test_that("ww_build_weights is stable", {
 })
 
 test_that("expected_errors", {
+  trip_dplyr_warning()
   expect_snapshot(
     ww_agreement_coefficient_vec(1:5, c(1:4, NA), na_action = c(na.omit, na.pass)),
     error = TRUE
