+++ DESCRIPTION
@@ -35,7 +35,7 @@ URL: https://github.com/mikemahoney218/waywiser,
     https://mikemahoney218.github.io/waywiser/
 BugReports: https://github.com/mikemahoney218/waywiser/issues
 Depends: 
-    R (>= 3.6)
+    R (>= 4.0)
 Imports: 
     dplyr,
     fields,
