### 2023-02-02 333cf42
Address reviewer comments (#17)
* Address reviewer comments

* Vignette doesn't need kableExtra anymore

* Trip dplyr lifecycle warning outside tests

* Bump min R version
