+++ inst/srr_template_spatial_yardstick.R
@@ -1,4 +1,5 @@
 test_that("srr: expected failures for {{{name}}}", {
+  trip_dplyr_warning()
   worldclim_predicted <- worldclim_simulation
   worldclim_predicted$predicted <- predict(
     lm(response ~ bio2 * bio10 * bio13 * bio19, data = worldclim_simulation),
@@ -226,6 +227,7 @@ test_that("srr: expected failures for {{{name}}}", {
 })
 
 test_that("other generic srr standards", {
+  trip_dplyr_warning()
   skip_if_not_installed("withr")
   worldclim_predicted <- worldclim_simulation
   worldclim_predicted$predicted <- predict(
