+++ tests/helper.R
@@ -1,3 +1,0 @@
-if (identical(Sys.getenv("NOT_CRAN"), "true")) {
-  Sys.setenv("waywiser_test_cast" = "true")
-}
