+++ R/global_geary.R
@@ -3,6 +3,10 @@
 #' Calculate the global Geary's C statistic for model residuals.
 #' `ww_global_geary_c()` returns the statistic itself, while
 #' `ww_global_geary_pvalue()` returns the associated p value.
+#' These functions are meant to help assess model predictions, for instance by
+#' identifying if there are clusters of higher residuals than expected. For
+#' statistical testing and inference applications, use
+#' [spdep::geary.test()] instead.
 #'
 #' These functions can be used for geographic or projected coordinate reference
 #' systems and expect 2D data.
