+++ R/local_getis.R
@@ -3,6 +3,9 @@
 #' Calculate the local Getis-Ord G and G* statistic for model residuals.
 #' `ww_local_getis_ord_g()` returns the statistic itself, while
 #' `ww_local_getis_ord_pvalue()` returns the associated p value.
+#' These functions are meant to help assess model predictions, for instance by
+#' identifying clusters of higher residuals than expected. For statistical
+#' testing and inference applications, use [spdep::localG_perm()] instead.
 #'
 #' These functions can be used for geographic or projected coordinate reference
 #' systems and expect 2D data.
