+++ pkgdown.yml
@@ -2,10 +2,11 @@ pandoc: 2.19.2
 pkgdown: 2.0.7
 pkgdown_sha: ~
 articles:
+  Feature-Matrix: Feature-Matrix.html
   multi-scale-assessment: multi-scale-assessment.html
   residual-autocorrelation: residual-autocorrelation.html
   waywiser: waywiser.html
-last_built: 2023-01-31T14:27Z
+last_built: 2023-02-02T20:19Z
 urls:
   reference: https://mikemahoney218.github.io/waywiser/reference
   article: https://mikemahoney218.github.io/waywiser/articles
