+++ vignettes/articles/Feature-Matrix.Rmd
@@ -19,18 +19,18 @@ package functions, check out the [package reference page](https://mikemahoney218
 funs <- tibble::tribble(
   ~ "Function", ~ "Description", ~ "Category", ~ "Reference",
   "ww_area_of_applicability()", "Calculate the 'area of applicability' of a model", "AOA", kableExtra::text_spec("H. Meyer and E. Pebesma. 2021. Predicting into unknown space? Estimating the area of applicability of spatial prediction models, Methods in Ecology and Evolution 12(9), pp 1620 - 1633, doi: 10.1111/2041-210X.13650.", link = "https://doi.org/10.1111/2041-210X.13650"),
-  "ww_multi_scale()", "Calculate model metrics for predictions aggregated to multiple levels", "Multi-scale assessment", kableExtra::text_spec("Riemann, R., Wilson, B. T., Lister, A., and Parks, S. (2010) An effective assessment protocol for continuous geospatial datasets of forest characteristics using USFS Forest Inventory and Analysis (FIA) data. Remote Sensing of Environment 114(10), pp 2337-2352, doi: 10.1016/j.rse.2010.05.010.", link = "https://doi.org/10.1016/j.rse.2010.05.010."),
-  "ww_global_moran_i() (plus _vec and _pvalue variants)", "Calculate Moran's I", "Spatial yardstick metrics", paste(kableExtra::text_spec("Moran, P.A.P. (1950). Notes on Continuous Stochastic Phenomena. Biometrika, 37(1/2), pp 17. doi: 10.2307/2332142", link = "https://doi.org/10.2307/2332142"), kableExtra::text_spec("Cliff, A. D., Ord, J. K. 1981 Spatial processes, Pion, p. 17.", link = "https://doi.org/10.1111/2041-210X.13650"), sep = "; "),
+  "ww_multi_scale()", "Calculate model metrics for predictions aggregated to multiple levels", "Multi-scale assessment", kableExtra::text_spec("Riemann, R., Wilson, B. T., Lister, A., and Parks, S. (2010) An effective assessment protocol for continuous geospatial datasets of forest characteristics using USFS Forest Inventory and Analysis (FIA) data. Remote Sensing of Environment 114(10), pp 2337-2352, doi: 10.1016/j.rse.2010.05.010", link = "https://doi.org/10.1016/j.rse.2010.05.010"),
+  "ww_global_moran_i() (plus _vec and _pvalue variants)", "Calculate Moran's I", "Spatial yardstick metrics", paste(kableExtra::text_spec("Moran, P.A.P. (1950). Notes on Continuous Stochastic Phenomena. Biometrika, 37(1/2), pp 17. doi: 10.2307/2332142", link = "https://doi.org/10.2307/2332142"), "Cliff, A. D., Ord, J. K. 1981 Spatial processes, Pion, p. 17.", sep = "; "),
[…]
