+++ pkgdown.yml
@@ -6,7 +6,7 @@ articles:
   multi-scale-assessment: multi-scale-assessment.html
   residual-autocorrelation: residual-autocorrelation.html
   waywiser: waywiser.html
-last_built: 2023-02-02T20:19Z
+last_built: 2023-02-02T20:40Z
 urls:
   reference: https://mikemahoney218.github.io/waywiser/reference
   article: https://mikemahoney218.github.io/waywiser/articles
