### 2023-02-02 #571 @mikemahoney218
Thank you for your comments, @Nowosad ! I believe I've addressed all your comments in the current development version of the package. I've responded to your points with a bit more detail below.

<hr />

> waywiser.Rmd: Have you considered explaining the example data first in this vignette?

Added! ([Commit](https://github.com/mikemahoney218/waywiser/pull/17/files#diff-65868d7690de5c91ac14638dfa2005364ab12268e42d7c8caafc00db630f0535R33-R36), [rendered](https://mikemahoney218.github.io/waywiser/articles/waywiser.html))

> waywiser.Rmd: Could you add a sentence or two better explaining where n came from in the Multi-scale model assessment section?; why is this a list (and also why cellsize is not a list in multi-scale-assessment.Rmd)?

Added! ([Commit](https://github.com/mikemahoney218/waywiser/pull/17/files#diff-65868d7690de5c91ac14638dfa2005364ab12268e42d7c8caafc00db630f0535R141-R151), [rendered](https://mikemahoney218.github.io/waywiser/articles/waywiser.html#multi-scale-model-assessment))

> waywiser.Rmd: Area of Applicability: could you expand the description of the importance argument in this vignette?

Added! ([Commit](https://github.com/mikemahoney218/waywiser/pull/17/files#diff-65868d7690de5c91ac14638dfa2005364ab12268e42d7c8caafc00db630f0535R182-R191), [rendered](https://mikemahoney218.github.io/waywiser/articles/waywiser.html#area-of-applicability))

> waywiser.Rmd: Area of Applicability: I would suggest also showing the result here (it is hard to think about an area of applicability without seeing it first)

Added! ([Commit](https://github.com/mikemahoney218/waywiser/pull/17/files#diff-65868d7690de5c91ac14638dfa2005364ab12268e42d7c8caafc00db630f0535R220-R231), [rendered](https://mikemahoney218.github.io/waywiser/articles/waywiser.html#area-of-applicability))

> waywiser.Rmd: I like the Feature Matrix table, however, I am not sure if it should be at the end of this vignette. Have you considered moving it to a standalone vignette (for better visibility)? Also, could you replace DOI codes with DOI urls?

I moved it to be a standalone article (so on the pkgdown site, but not built on CRAN), which also let me move kableExtra out of Suggests. ([Commit](https://github.com/mikemahoney218/waywiser/pull/17/files#diff-9a9075ac0a95ac4916417d3b03b9ad0791688a24ec78fd5020fe6a1defbb830f), [rendered](https://mikemahoney218.github.io/waywiser/articles/Feature-Matrix.html))

> residual-autocorrelation.Rmd and multi-scale-assessment.Rmd: What is the reason for using the %>% pipe here? Why not use the native pipe (|>)?

The main reason is because this package supports R >= 4, while the pipe was added in 4.1. Using the native pipe would make CI runs for R 4.0 fail, or only be run conditionally; in order for the vignettes to build on every supported version of R, I've kept the %>% pipe for now.

> residual-autocorrelation.Rmd: “This makes it easy to see what areas are poorly represented by our model” – could you elaborate on this sentence and explain which areas you are talking about?

I added a long discussion about what this means at the top of the vignette, and also elaborated a tiny bit at the end. ([Commit](https://github.com/mikemahoney218/waywiser/pull/17/files#diff-1051c2cdd1a164046f1b75caceb80ef87125bb8edf228e075731f9ebf2277191R17-R52), [rendered](https://mikemahoney218.github.io/waywiser/articles/residual-autocorrelation.html))

> GitHub Actions are mostly broken at the moment (I assume it is due to the recent dplyr changes)

These should be "fixed" for now, by tripping the new lifecycle warning in dplyr::summarise() at the top of each test file. The long-term fix will depend on if [yardstick changes to use `reframe()`](https://github.com/tidymodels/yardstick/issues/355).

> Regarding your “p-value” question: I think option 2 is fine.

Thank you! I've gone ahead and added this documentation throughout. ([Example of commit](https://github.com/mikemahoney218/waywiser/pull/17/files#diff-dfdf9c39baa8a7f9bda1d12409d3b1638150e802b3b6a20d217f33908d5efbf3R6-R8), [example of rendered](https://mikemahoney218.github.io/waywiser/reference/local_geary_c.html))

> Question: How do you see the future of this package? Do you plan to add any new features (e.g., spatial explainers)?

I'm not sure I know what spatial explainers are! That said, this is my broad vision for the package:

1. In the near term, I think I might extend `ww_multi_scale()` to accept raster inputs, for situations where you've predicted a large raster that won't fit entirely in memory as points. This is pretty much the only feature I have _planned_.
2. I'm open to adding yardstick metrics (such as `ww_agreement_coefficient`, `ww_local_geary_c`) if any are requested, or if I run into any in the literature. That said, I don't have any plans here, and don't know of any that would be useful to add. If the metric isn't coming from the spatial literature, it should probably live in yardstick instead.
3. My general goal is for waywiser to be a useful toolbox for assessing spatial models, and I view anything that falls under that headline as being "in scope". If an assessment method is coming from the spatial modeling world, then it's a good candidate for waywiser, even if it's not inherently spatial (so AOA, agreement coefficient, Willmott's D etc all fall under this). If it's not coming from the spatial modeling world, then I'd probably rather contribute techniques to [vip](https://koalaverse.github.io/vip/), [DALEX](https://modeloriented.github.io/DALEX/), [applicable](https://applicable.tidymodels.org/), or [yardstick](https://yardstick.tidymodels.org/).

With all that said, I'm not actively looking for things to add -- I'm currently only adding things that are useful for my own work. But if there are requests or PRs for other features, that's my basic outline for whether something belongs in waywiser or not.

Hope that answers the question!

> ☒ Should the author(s) deem it appropriate, I agree to be acknowledged as a package reviewer (“rev” role) in the package DESCRIPTION file.

Thanks! I'll add you as a reviewer in the DESCRIPTION once the package is accepted :smile:
