### 2023-02-02 8707fa3
Get rid of geometry column in a better way
