### 2023-02-03 7cd0c8a
Accept groups in multi-scale
