### 2023-02-04 #571 @Nowosad
@mikemahoney218 thanks for all of the improvements made.

> I'm not sure I know what spatial explainers are!

I've been thinking of something like https://geods.netlify.app/post/spatial-ml-model-diagnostics/.
