### 2023-02-04 #571 @mikemahoney218
@Nowosad I think that would be in scope for waywiser. I'm not planning on implementing it in the near future (I need to focus on my dissertation :sweat_smile:, so all the techniques I'm actively adding are things I'm going to use myself), but I could definitely see the package growing in that direction over time.
