### 2023-02-04 #571 @becarioprecario
@mikemahoney218<https://github.com/mikemahoney218> thanks for taking into account my comments. I hope that they have been useful.

I think these comments reflect that a few functions (namely, the p-value functions) are maybe a bit out of scope for the package, but I'm interested in what you (and others) think.

My opinion is that if a functions is not going to be used by the user it is best to keep it hidden. You can always use ::: if needed (for some reason…). So, the p-value functions is called from other functions in the package but I do not think that it is meant to be used directly by the user.

The p-value functions are included as "model assessment" tools because I've seen modeling projects use p-values to ID areas of concern, with regards to autocorrelation: locations with more extreme p-values for local autocorrelation metrics were selected for further investigation, to see if model specifications could be improved. In that sense, p-values are included as an assessment metric for predictive modeling, and not so much for statistical testing purposes. As such, waywiser lets you return p-values without also returning test statistics themselves, as this approach doesn't really require looking at the underlying test statistic values; extreme p-values are areas of interest, no matter what their actual statistic is.

I see. But p-values are tied to the stats statistics so I am not sure it is a good idea to only report p-values

Given all this, I see two desirable ways to address this set of comments:

  1.  Remove p-value functions from the package. This removes a use-case for the package (looking for extreme p-values to identify areas which might help improve model specifications), but also makes it more clear that the package is designed for assessing predictive accuracy, and is not oriented towards inference.
  2.  Retain p-value functions, but add documentation clarifying that for inference users are recommended to use spdep equivalents directly.

I would go for (2) and compute statistics and associated p-values from a main function. Then the user can decide on what to plot in a map.
