### 2023-02-06 #571 @mikemahoney218
@becarioprecario Thank you -- your comments have been extremely useful :smile:

My thinking (and experience) is that the p-value functions are called directly by users as a model diagnostic tool during the iteration process -- these p-values aren't being reported in a publication, but rather used to guide model development by highlighting hot-spots for model residuals (and hopefully helping to make a model misspecification clear, so it can be fixed _before_ any publication).

There's not really a great way for functions using the yardstick infrastructure to return two different statistics (so here, the test statistic and p-value). The idiomatic way to do so is to use `yardstick::metric_set()` to combine two functions (here, the test statistic and p-value functions), but that's something that's best left to the user, as metric sets can't be "expanded" to include additional metrics.

For example, if you want to calculate (for instance) global Moran's I with a p-value, plus an agreement coefficient, you can run `metrics <- yardstick::metric_set(ww_global_moran_i, ww_global_moran_p_value, ww_agreement_coefficient)` and then use the `metrics()` function with your data to get all three outputs. If waywiser provided a metric set (which is what functions like `ww_global_moran()` did, but note those functions were never in the submitted version of the package) then you couldn't call `yardstick::metric_set(ww_global_moran, ww_agreement_coefficient)`; you'd get an error.

That's why the "combined" functions were removed before this package was submitted; they don't work in a lot of places that users would expect them to be useful, and explaining the reason they work in a very different way than the rest of the package is pretty hard to communicate. Instead, all of the metrics provided by this package are pure yardstick metrics, without any of the weird edge cases. That means they're restricted to each returning a single type of statistic.

I've added documentation as described in (2) to these functions (see for instance https://github.com/mikemahoney218/waywiser/commit/333cf42cec3f6bddcd7f8b54150bc1f5dd8e365f#diff-45d2e91a37be2289564b4e1c987cbc8ac817ee874cc0ddcf19cfcdd8088c01feR6-R9 ). I could also add documentation about using `yardstick::metric_set()` to calculate both the test statistic and p-value at once, though I personally think it'd be better to not mention that; instead, the current documentation encourages people to use the spdep functions directly if they're looking to use p-values for other purposes than what I've described. This documentation (on using `metric_set()`) would probably look like the section on `metric_set()` that's in the [Getting Started vignette](https://mikemahoney218.github.io/waywiser/articles/waywiser.html#yardstick-extensions).

Alternatively, I could remove the p-value functions, if you think that having them at all without a combination function is harmful. But I don't think it makes sense to add combination functions; they introduce too many weird edge cases and don't idiomatically fit into yardstick.
