+++ README.md
@@ -1,194 +1,16 @@
 
 <!-- README.md is generated from README.Rmd. Please edit that file -->
 
-# waywiser <a href="https://mikemahoney218.github.io/waywiser/"><img src="man/figures/logo.png" align="right" height="138" /></a>
+# waywiserrr <a href="https://mikemahoney218.github.io/waywiser/"><img src="man/figures/logo.png" align="right" height="138" /></a>
 
-<!-- badges: start -->
+waywiserrr is
+[waywise**r**](https://mikemahoney218.github.io/waywiser/),
+**r**ewritten in **R**ust. I’m using this repo to experiment with using
+Rust in R packages, and to try and muddle my way through learning
+ndarray.
 
-[![R-CMD-check](https://github.com/mikemahoney218/waywiser/actions/workflows/R-CMD-check.yaml/badge.svg)](https://github.com/mikemahoney218/waywiser/actions/workflows/R-CMD-check.yaml)
-[![License:
-MIT](https://img.shields.io/badge/license-MIT-green)](https://choosealicense.com/licenses/mit/)
-[![Lifecycle:
-maturing](https://img.shields.io/badge/lifecycle-maturing-blue.svg)](https://lifecycle.r-lib.org/articles/stages.html#maturing)
-[![Project Status: Active – The project has reached a stable, usable
-state and is being actively
-developed.](https://www.repostatus.org/badges/latest/active.svg)](https://www.repostatus.org/#active)
-[![Codecov test
-coverage](https://codecov.io/gh/mikemahoney218/waywiser/branch/main/graph/badge.svg)](https://app.codecov.io/gh/mikemahoney218/waywiser?branch=main)
-[![CRAN
-status](https://www.r-pkg.org/badges/version/waywiser)](https://CRAN.R-project.org/package=waywiser)
-<!-- badges: end -->
-
-“Waywiser” is an old-timey name for a [surveyor’s
-wheel](https://en.wikipedia.org/wiki/Surveyor%27s_wheel), a device that
-makes measuring long distances easier than with measurement tools like a
-ruler or yardstick. The waywiser R package makes it easier to measure
-the performance of models fit to 2D spatial data by implementing a
-number of well-established assessment methods in a consistent, ergonomic
-toolbox; features include new
-[yardstick](https://yardstick.tidymodels.org/) metrics for measuring
-agreement and spatial autocorrelation, functions to assess model
-predictions across multiple scales, and methods to calculate the area of
-applicability of a model.
-
-## Installation
-
-You can install waywiser from CRAN via:
-
-``` r
-install.packages("waywiser")
-```
-
-You can install the development version of waywiser from
-[GitHub](https://github.com/) with:
-
-``` r
-# install.packages("devtools")
-remotes::install_github("mikemahoney218/waywiser")
-```
-
-## Example
-
-Let’s say that we fit a linear model predicting crimes against people as
-a function of literacy, using the `guerry` data included in waywiser:
-
-``` r
-library(waywiser)
-
-set.seed(123)
-split_idx <- sample(seq_len(nrow(guerry)), nrow(guerry) * 0.8)
-guerry_train <- guerry[split_idx, ]
-guerry_test <- guerry[-split_idx, ]
-
-crime_model <- lm(Crm_prs ~ Litercy, guerry_train)
-```
-
-We want to assess this model, to better understand how well it predicts
-crime rates across 1830s France. One method to do so is to evaluate our
-predictions at multiple levels of aggregation, as suggested by Riemann
-et al. (2010) (\<doi: 10.1016/j.rse.2010.05.010\>). This approach is
[…]

[…]
