+++ R/area_of_applicability.R
@@ -99,14 +99,13 @@
 #' must have a matching importance estimate, and all terms with importance
 #' estimates must be in the training data.
 #'
-#' @param ... Not currently used.
+#' @param na_rm A logical of length 1, indicating whether observations (in both
+#' training and testing) with `NA` values in predictors should be removed. Only
+#' predictor variables are considered, and this value has no impact on
+#' predictions (where `NA` values produce `NA` predictions). If `na_rm = FALSE`
+#' and `NA` values are found, this function returns an error.
 #'
-#' @param na_action A function which indicates what should happen when the data
-#' contain NAs. The default is `na.fail`; you may wish to set it to `na.omit`
-#' or any of the functions from the "zoo"
-#' package. This function ignores the value of `options("na.action")` in order
-#' to make cross-computer (and cross-session) results more stable. This argument
-#' must be of length 1; you cannot pass multiple functions to `na_action`.
+#' @param ... Not currently used.
 #'
 #' @details
 #'
@@ -171,11 +170,11 @@ ww_area_of_applicability.default <- function(x, ...) {
 
 #' @exportS3Method
 #' @rdname ww_area_of_applicability
-ww_area_of_applicability.data.frame <- function(x, testing = NULL, importance, ..., na_action = na.fail) {
+ww_area_of_applicability.data.frame <- function(x, testing = NULL, importance, ..., na_rm = FALSE) {
   rlang::check_dots_empty()
   training <- hardhat::mold(x, NA_real_)
   if (!is.null(testing)) testing <- hardhat::mold(testing, NA_real_)
-  create_aoa(training, testing, importance, ..., na_action = na_action)
+  create_aoa(training, testing, importance, ..., na_rm = na_rm)
 }
 
 #' @exportS3Method
@@ -184,7 +183,7 @@ ww_area_of_applicability.matrix <- ww_area_of_applicability.data.frame
 
 #' @exportS3Method
 #' @rdname ww_area_of_applicability
-ww_area_of_applicability.formula <- function(x, data, testing = NULL, importance, ..., na_action = na.fail) {
+ww_area_of_applicability.formula <- function(x, data, testing = NULL, importance, ..., na_rm = FALSE) {
   rlang::check_dots_empty()
 
   # This method is also used for recipes, which don't need blueprints:
@@ -217,7 +216,7 @@ ww_area_of_applicability.formula <- function(x, data, testing = NULL, importance
     )
   }
 
-  create_aoa(training, processed_testing, importance, ..., na_action = na_action)
+  create_aoa(training, processed_testing, importance, ..., na_rm = na_rm)
 }
 
 #' @exportS3Method
@@ -226,7 +225,7 @@ ww_area_of_applicability.recipe <- ww_area_of_applicability.formula
 
 #' @exportS3Method
 #' @rdname ww_area_of_applicability
-ww_area_of_applicability.rset <- function(x, y = NULL, importance, ..., na_action = na.fail) {
+ww_area_of_applicability.rset <- function(x, y = NULL, importance, ..., na_rm = FALSE) {
   rlang::check_dots_empty()
 
   if (missing(y) || identical(y, NULL) || identical(y, NA)) y <- NA_real_
@@ -249,26 +248,19 @@ ww_area_of_applicability.rset <- function(x, y = NULL, importance, ..., na_actio
         testing,
         importance,
         include_di = TRUE,
-        na_action = na_action
+        na_rm = na_rm
       )
     }
   )
 
   aoa <- aoa_calcs[[1]]
   training <- if (identical(y, NA_real_)) {
-    check_for_missing(
[…]
