### 2023-02-15 3dd90d0
Add terra support in ww_multi_scale()
