+++ tests/testthat/test-multi_scale.R
@@ -353,3 +353,80 @@ test_that("other generic srr standards", {
   )
 })
 
+test_that("raster method works", {
+  skip_if_not_installed("terra")
+  skip_if_not_installed("exactextractr")
+  skip_if_not_installed("withr")
+
+  r1 <- matrix(nrow = 10, ncol = 10)
+  r1[] <- 1
+  r1 <- terra::rast(r1)
+
+  r2 <- matrix(nrow = 10, ncol = 10)
+  r2[] <- 2
+  r2 <- terra::rast(r2)
+
+  expect_identical(
+    ww_multi_scale(truth = r1, estimate = r2, metrics = yardstick::rmse, n = 1)$.estimate,
+    1
+  )
+
+  expect_identical(
+    ww_multi_scale(truth = r1, estimate = r1, metrics = yardstick::rmse, n = 1)$.estimate,
+    0
+  )
+})
+
+test_that("raster method is equivalent", {
+  skip_if_not_installed("terra")
+  skip_if_not_installed("exactextractr")
+  skip_if_not_installed("withr")
+
+  withr::with_seed(
+    123,
+    {
+      x <- terra::rast(matrix(rnorm(100), 10))
+      y <- terra::rast(matrix(rnorm(100), 10))
+    }
+  )
+
+  z <- sf::st_as_sf(terra::as.points(c(x, y)))
+
+  sf_method <- ww_multi_scale(z, lyr.1, lyr.1.1, n = 2)
+  sf::st_geometry(sf_method$.grid[[1]]) <- NULL
+  sf::st_geometry(sf_method$.grid[[2]]) <- NULL
+
+  terra_method <- ww_multi_scale(truth = x, estimate = y, n = 2)
+  sf::st_geometry(terra_method$.grid[[1]]) <- NULL
+  sf::st_geometry(terra_method$.grid[[2]]) <- NULL
+
+  expect_identical(sf_method, terra_method, tolerance = 1e-6)
+
+})
+
+test_that("raster method errors as expected", {
+  skip_if_not_installed("terra")
+  skip_if_not_installed("exactextractr")
+
+  r1 <- matrix(nrow = 10, ncol = 10)
+
+  expect_snapshot(
+    ww_multi_scale(truth = 1),
+    error = TRUE
+  )
+  expect_snapshot(
+    ww_multi_scale(truth = c(terra::rast(r1), terra::rast(r1))),
+    error = TRUE
+  )
+
+  expect_snapshot(
+    ww_multi_scale(truth = terra::rast(r1), estimate = 1),
+    error = TRUE
+  )
+  expect_snapshot(
+    ww_multi_scale(truth = terra::rast(r1), estimate = c(terra::rast(r1), terra::rast(r1))),
+    error = TRUE
+  )
+
+
+})
