+++ R/waywiser-package.R
@@ -5,5 +5,6 @@
 #' @importFrom rlang enquo .data
 #' @importFrom stats predict complete.cases na.fail
 #' @importFrom yardstick new_numeric_metric
+utils::globalVariables(c(".truth", ".estimate"))
 ## usethis namespace: end
 NULL
