+++ R/multi_scale.R
@@ -1,6 +1,11 @@
 #' Evaluate metrics at multiple scales of aggregation
 #'
 #' @details
+#'
+#' If `data` is `NULL`, then `truth` and `estimate` should both be `SpatRaster`
+#' objects, as created via [terra::rast()]. These rasters will then be
+#' aggregated to each grid using [exactextractr::exact_extract()].
+#'
 #' The grid blocks can be controlled by passing arguments to
 #' [sf::st_make_grid()] via `...`. Some particularly useful arguments include:
 #'
@@ -22,9 +27,13 @@
 #' This function can be used for geographic or projected coordinate reference
 #' systems and expect 2D data.
 #'
-#' @param data A point geometry `sf` object containing the columns specified by
-#' the truth and estimate arguments.
-#' @inheritParams yardstick::rmse
+#' @param data Either a point geometry `sf` object containing the columns
+#' specified by the `truth` and `estimate` arguments, or `NULL` if `truth` and
+#' `estimate` are `SpatRaster` objects.
+#' @param truth,estimate If `data` is an `sf` object, the names (optionally
+#' unquoted) for the columns in `data` containing the true and predicted values,
+#' respectively. If `data` is `NULL`, `SpatRaster` objects with a single layer
+#' containing the true and predicted values, respectively.
 #' @param metrics Either a [yardstick::metric_set()] object, or a list of
 #' functions which will be used to construct a [yardstick::metric_set()] object
 #' specifying the performance metrics to evaluate at each scale.
@@ -35,12 +44,18 @@
 #' For instance, passing `n = list(c(1, 2))` will create a single 1x2 grid;
 #' passing `n = c(1, 2)` will create a 1x1 grid _and_ a 2x2 grid.
 #' @param aggregation_function The function to use to aggregate predictions and
-#' true values at various scales, by default [mean()]. You can pass any function
-#' which takes a single vector and returns a scalar.
+#' true values at various scales, by default [mean()]. For the `sf` method,
+#' you can pass any function which takes a single vector and returns a scalar.
+#' For raster methods, any function accepted by
+#' [exactextractr::exact_extract()] (note that built-in function names must be
+#' quoted).
 #' @param autoexpand_grid Boolean: if `data` is in geographic coordinates and
 #' `grids` aren't provided, the grids generated by [sf::st_make_grid()] may not
 #' contain all observations. If `TRUE`, this function will automatically expand
 #' generated grids by a tiny factor to attempt to capture all observations.
+#' @param progress Boolean: if `data` is `NULL`, should aggregation via
+#' [exactextractr::exact_extract()] show a progress bar? Separate progress bars
+#' will be shown for each time `truth` and `estimate` are aggregated.
 #' @inheritParams ww_area_of_applicability
 #'
 #' @return A tibble with six columns: `.metric`, with the name
@@ -85,6 +100,104 @@
 #'
 #' @export
 ww_multi_scale <- function(
+    data = NULL,
+    truth,
+    estimate,
+    metrics = list(yardstick::rmse, yardstick::mae),
+    grids = NULL,
+    ...,
+    na_rm = TRUE,
+    aggregation_function = "mean",
+    autoexpand_grid = TRUE,
+    progress = TRUE
+) {
[…]
