+++ man/ww_multi_scale.Rd
@@ -5,31 +5,27 @@
 \title{Evaluate metrics at multiple scales of aggregation}
 \usage{
 ww_multi_scale(
-  data,
+  data = NULL,
   truth,
   estimate,
   metrics = list(yardstick::rmse, yardstick::mae),
   grids = NULL,
   ...,
   na_rm = TRUE,
-  aggregation_function = mean,
-  autoexpand_grid = TRUE
+  aggregation_function = "mean",
+  autoexpand_grid = TRUE,
+  progress = TRUE
 )
 }
 \arguments{
-\item{data}{A point geometry `sf` object containing the columns specified by
-the truth and estimate arguments.}
+\item{data}{Either a point geometry `sf` object containing the columns
+specified by the `truth` and `estimate` arguments, or `NULL` if `truth` and
+`estimate` are `SpatRaster` objects.}
 
-\item{truth}{The column identifier for the true results
-(that is \code{numeric}). This should be an unquoted column name although
-this argument is passed by expression and supports
-\link[rlang:topic-inject]{quasiquotation} (you can unquote column
-names). For \verb{_vec()} functions, a \code{numeric} vector.}
-
-\item{estimate}{The column identifier for the predicted
-results (that is also \code{numeric}). As with \code{truth} this can be
-specified different ways but the primary method is to use an
-unquoted variable name. For \verb{_vec()} functions, a \code{numeric} vector.}
+\item{truth, estimate}{If `data` is an `sf` object, the names (optionally
+unquoted) for the columns in `data` containing the true and predicted values,
+respectively. If `data` is `NULL`, `SpatRaster` objects with a single layer
+containing the true and predicted values, respectively.}
 
 \item{metrics}{Either a [yardstick::metric_set()] object, or a list of
 functions which will be used to construct a [yardstick::metric_set()] object
@@ -43,17 +39,27 @@ specifying polygon boundaries to use for assessments.}
 For instance, passing `n = list(c(1, 2))` will create a single 1x2 grid;
 passing `n = c(1, 2)` will create a 1x1 grid _and_ a 2x2 grid.}
 
-\item{na_rm}{A \code{logical} value indicating whether \code{NA}
-values should be stripped before the computation proceeds.}
+\item{na_rm}{A logical of length 1, indicating whether observations (in both
+training and testing) with `NA` values in predictors should be removed. Only
+predictor variables are considered, and this value has no impact on
+predictions (where `NA` values produce `NA` predictions). If `na_rm = FALSE`
+and `NA` values are found, this function returns an error.}
 
 \item{aggregation_function}{The function to use to aggregate predictions and
-true values at various scales, by default [mean()]. You can pass any function
-which takes a single vector and returns a scalar.}
+true values at various scales, by default [mean()]. For the `sf` method,
+you can pass any function which takes a single vector and returns a scalar.
+For raster methods, any function accepted by
+[exactextractr::exact_extract()] (note that built-in function names must be
+quoted).}
 
 \item{autoexpand_grid}{Boolean: if `data` is in geographic coordinates and
 `grids` aren't provided, the grids generated by [sf::st_make_grid()] may not
 contain all observations. If `TRUE`, this function will automatically expand
 generated grids by a tiny factor to attempt to capture all observations.}
+
+\item{progress}{Boolean: if `data` is `NULL`, should aggregation via
+[exactextractr::exact_extract()] show a progress bar? Separate progress bars
+will be shown for each time `truth` and `estimate` are aggregated.}
 }
 \value{
 A tibble with six columns: `.metric`, with the name
@@ -68,6 +74,10 @@ a given assessment.
 Evaluate metrics at multiple scales of aggregation
 }
 \details{
[…]

[…]
