+++ man/ww_multi_scale.Rd
@@ -67,8 +67,8 @@ of the metric that the row describes; `.estimator`, with the name of the
 estimator used, `.estimate`, with the output of the metric function;
 `.grid_args`, with the arguments passed to [sf::st_make_grid()] via `...`
 (if any), `.grid`, containing the grids used to aggregate predictions,
-and `.notes`, which will indicate any observations which were not used in
-a given assessment.
+and `.notes`, which (if `data` is an `sf` object) will indicate any
+observations which were not used in a given assessment.
 }
 \description{
 Evaluate metrics at multiple scales of aggregation
