+++ R/multi_scale.R
@@ -1,10 +1,17 @@
 #' Evaluate metrics at multiple scales of aggregation
 #'
-#' @details
+#' @section Raster inputs:
 #'
 #' If `data` is `NULL`, then `truth` and `estimate` should both be `SpatRaster`
 #' objects, as created via [terra::rast()]. These rasters will then be
 #' aggregated to each grid using [exactextractr::exact_extract()].
+#'
+#' Grids are calculated using the bounding box of `truth`, under the assumption
+#' that you may have extrapolated into regions which do not have matching "true"
+#' values. This function does not check that `truth` and `estimate` overlap at
+#' all, or that they are at all contained within the grid.
+#'
+#' @section Creating grid blocks:
 #'
 #' The grid blocks can be controlled by passing arguments to
 #' [sf::st_make_grid()] via `...`. Some particularly useful arguments include:
@@ -25,7 +32,7 @@
 #' number of cells.
 #'
 #' This function can be used for geographic or projected coordinate reference
-#' systems and expect 2D data.
+#' systems and expects 2D data.
 #'
 #' @param data Either a point geometry `sf` object containing the columns
 #' specified by the `truth` and `estimate` arguments, or `NULL` if `truth` and
@@ -48,7 +55,13 @@
 #' you can pass any function which takes a single vector and returns a scalar.
 #' For raster methods, any function accepted by
 #' [exactextractr::exact_extract()] (note that built-in function names must be
-#' quoted).
+#' quoted). Note that this function does _not_ pay attention to the value of
+#' `na_rm`; any NA handling you want to do during aggregation should be handled
+#' by this function directly.
+#' @param na_rm Boolean: Should polygons with NA values be removed before
+#' calculating metrics? Note that this does _not_ impact how values are
+#' aggregated to polygons: if you want to remove NA values before aggregating,
+#' provide a function to `aggregation_function` which will remove NA values.
 #' @param autoexpand_grid Boolean: if `data` is in geographic coordinates and
 #' `grids` aren't provided, the grids generated by [sf::st_make_grid()] may not
 #' contain all observations. If `TRUE`, this function will automatically expand
@@ -56,15 +69,16 @@
 #' @param progress Boolean: if `data` is `NULL`, should aggregation via
 #' [exactextractr::exact_extract()] show a progress bar? Separate progress bars
 #' will be shown for each time `truth` and `estimate` are aggregated.
-#' @inheritParams ww_area_of_applicability
 #'
 #' @return A tibble with six columns: `.metric`, with the name
 #' of the metric that the row describes; `.estimator`, with the name of the
 #' estimator used, `.estimate`, with the output of the metric function;
 #' `.grid_args`, with the arguments passed to [sf::st_make_grid()] via `...`
 #' (if any), `.grid`, containing the grids used to aggregate predictions,
-#' and `.notes`, which (if `data` is an `sf` object) will indicate any
-#' observations which were not used in a given assessment.
+#' as well as the aggregated values of `truth` and `estimate` as well as the
+#' count of non-NA values for each, and `.notes`, which (if `data` is an `sf`
+#' object) will indicate any observations which were not used in a given
+#' assessment.
 #'
 #' @examplesIf rlang::is_installed("modeldata")
 #' data(ames, package = "modeldata")
@@ -147,6 +161,10 @@ ww_multi_scale.SpatRaster <- function(
     rlang::abort("`estimate` must be a SpatRaster with only one layer.")
   }
 
+  if (sf::st_crs(truth) != sf::st_crs(estimate)) {
+    rlang::abort("`truth` and `estimate` must share a CRS.")
+  }
+
   metrics <- handle_metrics(metrics)
[…]
