+++ tests/testthat/_snaps/multi_scale.md
@@ -6,12 +6,12 @@
       # A tibble: 6 x 6
         .metric .estimator .estimate .grid_args       .grid          .notes          
         <chr>   <chr>          <dbl> <list>           <list>         <list>          
-      1 rmse    standard      75108. <tibble [1 x 2]> <sf [344 x 3]> <tibble [0 x 2]>
-      2 mae     standard      54833. <tibble [1 x 2]> <sf [344 x 3]> <tibble [0 x 2]>
-      3 rmse    standard      82076. <tibble [1 x 2]> <sf [103 x 3]> <tibble [0 x 2]>
-      4 mae     standard      60467. <tibble [1 x 2]> <sf [103 x 3]> <tibble [0 x 2]>
-      5 rmse    standard      27862. <tibble [1 x 2]> <sf [5 x 3]>   <tibble [0 x 2]>
-      6 mae     standard      23267. <tibble [1 x 2]> <sf [5 x 3]>   <tibble [0 x 2]>
+      1 rmse    standard      75108. <tibble [1 x 2]> <sf [344 x 5]> <tibble [0 x 2]>
+      2 mae     standard      54833. <tibble [1 x 2]> <sf [344 x 5]> <tibble [0 x 2]>
+      3 rmse    standard      82076. <tibble [1 x 2]> <sf [103 x 5]> <tibble [0 x 2]>
+      4 mae     standard      60467. <tibble [1 x 2]> <sf [103 x 5]> <tibble [0 x 2]>
+      5 rmse    standard      27862. <tibble [1 x 2]> <sf [5 x 5]>   <tibble [0 x 2]>
+      6 mae     standard      23267. <tibble [1 x 2]> <sf [5 x 5]>   <tibble [0 x 2]>
 
 ---
 
@@ -21,7 +21,7 @@
       # A tibble: 1 x 6
         .metric .estimator .estimate .grid_args       .grid          .notes          
         <chr>   <chr>          <dbl> <list>           <list>         <list>          
-      1 rmse    standard      75108. <tibble [0 x 0]> <sf [344 x 3]> <tibble [0 x 2]>
+      1 rmse    standard      75108. <tibble [0 x 0]> <sf [344 x 5]> <tibble [0 x 2]>
 
 ---
 
@@ -35,8 +35,8 @@
       # A tibble: 2 x 6
         .metric .estimator .estimate .grid_args       .grid        .notes          
         <chr>   <chr>          <dbl> <list>           <list>       <list>          
-      1 rmse    standard        51.1 <tibble [1 x 1]> <sf [1 x 3]> <tibble [1 x 2]>
-      2 mae     standard        51.1 <tibble [1 x 1]> <sf [1 x 3]> <tibble [1 x 2]>
+      1 rmse    standard        51.1 <tibble [1 x 1]> <sf [1 x 5]> <tibble [1 x 2]>
+      2 mae     standard        51.1 <tibble [1 x 1]> <sf [1 x 5]> <tibble [1 x 2]>
 
 # expected errors
 
@@ -98,8 +98,8 @@
       # A tibble: 2 x 6
         .metric .estimator .estimate .grid_args       .grid        .notes          
         <chr>   <chr>          <dbl> <list>           <list>       <list>          
-      1 rmse    standard     0.00739 <tibble [1 x 1]> <sf [4 x 3]> <tibble [0 x 2]>
-      2 mae     standard     0.00691 <tibble [1 x 1]> <sf [4 x 3]> <tibble [0 x 2]>
+      1 rmse    standard     0.00739 <tibble [1 x 1]> <sf [4 x 5]> <tibble [0 x 2]>
[…]
