+++ NAMESPACE
@@ -19,7 +19,7 @@ S3method(ww_local_getis_ord_g,data.frame)
 S3method(ww_local_getis_ord_g_pvalue,data.frame)
 S3method(ww_local_moran_i,data.frame)
 S3method(ww_local_moran_pvalue,data.frame)
-S3method(ww_multi_scale,NULL)
+S3method(ww_multi_scale,SpatRaster)
 S3method(ww_multi_scale,sf)
 S3method(ww_systematic_agreement_coefficient,data.frame)
 S3method(ww_systematic_mpd,data.frame)

[…]
