### 2023-02-21 #571 @mikemahoney218
@becarioprecario I added a summary of this discussion to the "residual autocorrelation" vignette:
https://github.com/mikemahoney218/waywiser/commit/c0860a1b974b9b0e2b5d55b990657d8a7706fb28

Thank you again for your time reviewing this package, it's been a real help.
