+++ DESCRIPTION
@@ -7,6 +7,10 @@ Authors@R: c(
            comment = c(ORCID = "0000-0003-2402-304X")),
     person("Lucas", "Johnson", , "lucas.k.johnson03@gmail.com", role = c("ctb"),
            comment = c(ORCID = "0000-0002-7953-0260")),
+    person("Virgilio", "Gómez-Rubio", role = "rev",
+    comment = "Virgilio reviewed the package (v. 0.2.0.9000) for rOpenSci, see <https://github.com/ropensci/software-review/issues/571>"),
+    person("Jakub", "Nowosad", role = "rev",
+    comment = "Jakub reviewed the package (v. 0.2.0.9000) for rOpenSci, see <https://github.com/ropensci/software-review/issues/571>"),
     person("RStudio", role = c("cph", "fnd"))
   )
 Description: Assessing predictive models of spatial data can be challenging, 
