### 2023-02-27 #571 @mikemahoney218
@Paula-Moraga / @maelle Is there any chance someone could invite me to @ropensci/waywiser manually? For whatever reason, the automated process has never worked for me ( https://github.com/ropensci/software-review/issues/416 , https://github.com/ropensci/software-review/issues/521 ). I don't _think_ I've altered any account settings to block invites, but for some reason it's needed to be done manually every time.
